# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TradeXV2 API",
    "description": "Quantitative Trading Platform API - Market Data, Analytics, Scanner, Strategy, Replay, Backtesting",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/market/candles": {
      "get": {
        "tags": [
          "Market Data"
        ],
        "summary": "Get Candles",
        "description": "Get historical OHLCV candles from the data lake.\n\nSupports multiple timeframes and date range filtering.\nData is sourced from Parquet files in market_data/.\n\nCache-Control headers vary by timeframe:\n- 1m-5m: max-age=30 (30 seconds)\n- 15m-4h: max-age=300 (5 minutes)\n- 1d-1w: max-age=3600 (1 hour)\n\nFreshness headers:\n- X-Data-Stale: true/false\n- X-Data-Last-Update: date of most recent candle\n- X-Data-Days-Old: age in days\n- X-Data-Freshness: ISO timestamp of most recent candle\n- X-Data-Type: historical",
        "operationId": "get_candles_api_v1_market_candles_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Symbol to fetch",
              "title": "Symbol"
            },
            "description": "Symbol to fetch"
          },
          {
            "name": "exchange",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Exchange (NSE, BSE, NFO, etc.)",
              "default": "NSE",
              "title": "Exchange"
            },
            "description": "Exchange (NSE, BSE, NFO, etc.)"
          },
          {
            "name": "timeframe",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Timeframe (1m, 3m, 5m, 15m, 30m, 1h, 4h, 1d, 1w)",
              "title": "Timeframe"
            },
            "description": "Timeframe (1m, 3m, 5m, 15m, 30m, 1h, 4h, 1d, 1w)"
          },
          {
            "name": "from_ts",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start timestamp (ms)",
              "title": "From Ts"
            },
            "description": "Start timestamp (ms)"
          },
          {
            "name": "to_ts",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End timestamp (ms)",
              "title": "To Ts"
            },
            "description": "End timestamp (ms)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "description": "Max candles",
              "default": 200,
              "title": "Limit"
            },
            "description": "Max candles"
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandlesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/market/quote/{symbol}": {
      "get": {
        "tags": [
          "Market Data"
        ],
        "summary": "Get Quote",
        "description": "Get latest quote/LTP snapshot for a symbol via domain objects.\n\nCache-Control: max-age=10 (10 seconds) - quotes change frequently.",
        "operationId": "get_quote_api_v1_market_quote__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "exchange",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Exchange",
              "default": "NSE",
              "title": "Exchange"
            },
            "description": "Exchange"
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/portfolio/positions": {
      "get": {
        "tags": [
          "Portfolio"
        ],
        "summary": "Get Positions",
        "description": "Get current positions from OMS.\n\nReturns open and closed positions from broker or paper trading.",
        "operationId": "get_positions_api_v1_portfolio_positions_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter: open, closed, all",
              "title": "Status"
            },
            "description": "Filter: open, closed, all"
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolio/holdings": {
      "get": {
        "tags": [
          "Portfolio"
        ],
        "summary": "Get Holdings",
        "description": "Get current holdings/portfolio.\n\nReturns long-term holdings with cost basis and current value.\nUses real PositionManager data, filtering for delivery positions.",
        "operationId": "get_holdings_api_v1_portfolio_holdings_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Candle": {
        "properties": {
          "t": {
            "type": "integer",
            "title": "T",
            "description": "Timestamp (ms)"
          },
          "o": {
            "type": "number",
            "title": "O",
            "description": "Open"
          },
          "h": {
            "type": "number",
            "title": "H",
            "description": "High"
          },
          "l": {
            "type": "number",
            "title": "L",
            "description": "Low"
          },
          "c": {
            "type": "number",
            "title": "C",
            "description": "Close"
          },
          "v": {
            "type": "number",
            "title": "V",
            "description": "Volume"
          },
          "oi": {
            "type": "number",
            "title": "Oi",
            "description": "Open interest",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "t",
          "o",
          "h",
          "l",
          "c",
          "v"
        ],
        "title": "Candle",
        "description": "OHLCV candle."
      },
      "CandlesResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange",
            "default": "NSE"
          },
          "candles": {
            "items": {
              "$ref": "#/components/schemas/Candle"
            },
            "type": "array",
            "title": "Candles"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "timeframe",
          "candles",
          "count"
        ],
        "title": "CandlesResponse",
        "description": "Candle data response."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Holding": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "average_price": {
            "type": "number",
            "title": "Average Price"
          },
          "current_price": {
            "type": "number",
            "title": "Current Price"
          },
          "invested_value": {
            "type": "number",
            "title": "Invested Value"
          },
          "current_value": {
            "type": "number",
            "title": "Current Value"
          },
          "pnl": {
            "type": "number",
            "title": "Pnl"
          },
          "pnl_percent": {
            "type": "number",
            "title": "Pnl Percent"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "exchange",
          "quantity",
          "average_price",
          "current_price",
          "invested_value",
          "current_value",
          "pnl",
          "pnl_percent"
        ],
        "title": "Holding",
        "description": "Holding model."
      },
      "HoldingsResponse": {
        "properties": {
          "holdings": {
            "items": {
              "$ref": "#/components/schemas/Holding"
            },
            "type": "array",
            "title": "Holdings"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "total_value": {
            "type": "number",
            "title": "Total Value"
          },
          "total_invested": {
            "type": "number",
            "title": "Total Invested"
          },
          "total_pnl": {
            "type": "number",
            "title": "Total Pnl"
          }
        },
        "type": "object",
        "required": [
          "holdings",
          "count",
          "total_value",
          "total_invested",
          "total_pnl"
        ],
        "title": "HoldingsResponse",
        "description": "All holdings response."
      },
...
      "Position": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "average_price": {
            "type": "number",
            "title": "Average Price"
          },
          "current_price": {
            "type": "number",
            "title": "Current Price"
          },
          "unrealized_pnl": {
            "type": "number",
            "title": "Unrealized Pnl"
          },
          "realized_pnl": {
            "type": "number",
            "title": "Realized Pnl"
          },
          "pnl_pct": {
            "type": "number",
            "title": "Pnl Pct"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "exchange",
          "quantity",
          "average_price",
          "current_price",
          "unrealized_pnl",
          "realized_pnl",
          "pnl_pct"
        ],
        "title": "Position",
        "description": "Simplified position model."
      },
      "PositionsResponse": {
        "properties": {
          "positions": {
            "items": {
              "$ref": "#/components/schemas/Position"
            },
            "type": "array",
            "title": "Positions"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "total_pnl": {
            "type": "number",
            "title": "Total Pnl"
          },
          "total_pnl_percent": {
            "type": "number",
            "title": "Total Pnl Percent"
          }
        },
        "type": "object",
        "required": [
          "positions",
          "count",
          "total_pnl",
          "total_pnl_percent"
        ],
        "title": "PositionsResponse",
        "description": "All positions response."
      },
      "QuoteResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange"
          },
          "ltp": {
            "type": "number",
            "title": "Ltp"
          },
          "timestamp": {
            "type": "integer",
            "title": "Timestamp",
            "description": "Timestamp (ms)"
          },
          "bid": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bid"
          },
          "ask": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ask"
          },
          "bid_qty": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bid Qty"
          },
          "ask_qty": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ask Qty"
          },
          "volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          },
          "oi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oi"
          },
          "open": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open"
          },
          "high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "High"
          },
          "low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low"
          },
          "close": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "exchange",
          "ltp",
          "timestamp"
        ],
        "title": "QuoteResponse",
        "description": "Latest quote/LTP snapshot."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}